Fix `\parindent`
Even though the document settings were meant to set `\parindent` to 1.5em, somewhere else it was getting forced to 0pt.  I removed those places where I could find that statement explicitly, but it’s still happening somewhere implicitly so I’ve changed the default setting so that it is applied at the end of the class.  This allows the default to take effect while still permitting the user to manipulate it in the preamble should they so desire.