Shift capitalization enforcement for unnumbered chapter headings
Instead of enforcing the capitalization of chapter headings for each unnumbered chapter individually, this puts that responsibility on the command which typesets the heading.  The result is that the enforcement is more uniform and more robust to changes in header names (as might happen when `babel` or a similar package is used).