Change capitalization in author bio
Whittier indicated Graduate School asked that "The University of Maine" be "the University of Maine" in the automatically generated sentence in the author bio.  This is in violation of what's in the guidelines and so I won't merge this into master until I receive confirmation from the Graduate School directly.